수정

+++ include/linux/gps.h
@@ -11,7 +11,8 @@ struct gps_location {
 	float accuracy; /* in meters */
 };
 
-asmlinkage int set_gps_location(struct gps_location __user *loc);
-asmlinkage int get_gps_location(const char *pathname, struct gps_location __user *loc);
-
+/*
+asmlinkage long set_gps_location(struct gps_location __user *loc);
+asmlinkage long get_gps_location(const char *pathname, struct gps_location __user *loc);
+*/
 #endif
